refactor: Create uhk-header component
Create a header wrapper component that will contains the undo
notification component.
