+++ shared/src/components/macro/not-found/macro-not-found.component.html
@@ -1,3 +1,4 @@
+<uhk-header></uhk-header>
 <div class="not-found">
     You don't have any macros. Try to add one!
 </div>
