+++ shared/src/components/uhk-header/uhk-header.html
@@ -1,0 +1,3 @@
+<div>
+    <ng-content></ng-content>
+</div>
