+++ shared/src/components/uhk-header/uhk-header.ts
@@ -1,0 +1,9 @@
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+
+@Component({
+    selector: 'uhk-header',
+    templateUrl: './uhk-header.html',
+    changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class UhkHeader {
+}
